fixing extension loading error
