+++ Project.toml
@@ -16,10 +16,10 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [weakdeps]
-Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
+EnzymeCore = "f151be2c-9106-41f4-ab19-57ee4f262869"
 
 [extensions]
-NormalizingFlowsEnzymeExt = "Enzyme"
+NormalizingFlowsEnzymeCoreExt = "EnzymeCore"
 
 [compat]
 ADTypes = "1"
@@ -32,3 +32,6 @@ ProgressMeter = "1.0.0"
 Requires = "1"
 StatsBase = "0.33, 0.34"
 julia = "1.10"
+
+[extras]
+EnzymeCore = "f151be2c-9106-41f4-ab19-57ee4f262869"
