+++ src/NormalizingFlows.jl
@@ -79,19 +79,17 @@ include("optimize.jl")
 include("objectives.jl")
 
 
-# optional dependencies 
-if !isdefined(Base, :get_extension) # check whether :get_extension is defined in Base
+if !isdefined(Base, :get_extension)
     using Requires
 end
 
-# Question: should Exts be loaded here or in train.jl? 
-function __init__()
-    @static if !isdefined(Base, :get_extension)
-        @require Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9" include(
-            "../ext/NormalizingFlowsEnzymeExt.jl"
+
+@static if !isdefined(Base, :get_extension)
+    function __init__()
+        @require EnzymeCore = "f151be2c-9106-41f4-ab19-57ee4f262869" include(
+            joinpath(@__DIR__, "../ext/NormalizingFlowsEnzymeCoreExt.jl")
         )
     end
 end
-
 
 end
